+++ CHANGELOG.md
@@ -8,6 +8,8 @@ v0.9.0
 
   * Output path (screenshots, videos, visualtests) can be specified with a system property: *Files_output_path*
   * Update app_strings in Appium tests only if the driver has changed
+  * With a visual assertion error, the test can fail or give an error message and continue
+    New config property 'visualtests_fail' in [Server] section to fail the test when there is a visual error
 
 v0.8.6
 ------